ADD optic code improvementz
