+++ q42stats/src/main/java/com/q42/q42stats/library/collector/GooglePayCollector.kt
@@ -1,28 +1,24 @@
 package com.q42.q42stats.library.collector
 
 import android.content.Context
-import com.q42.q42stats.library.pay.GooglePayConfiguration.cardNetworks
 import com.q42.q42stats.library.pay.GooglePayManager
-import com.q42.q42stats.library.pay.GooglePayManager.IsReadyToPayResponse
+import com.q42.q42stats.library.pay.cardNetworks
+import com.q42.q42stats.library.pay.isGooglePlayEnabled
 import java.io.Serializable
 
-/** Collects Google Pay settings of the user */
+/** Collects Google Pay settings of the user
+ *
+ *  NOTE: The Google Pay "add payment method" does not support PayPal in NL atm, so we don't check
+ *  for PayPal. We need to be a PayPal merchant to be able to do so.
+ */
 internal object GooglePayCollector {
 
     suspend fun collect(context: Context) = mutableMapOf<String, Serializable>().apply {
 
         val googlePayManager = GooglePayManager(context)
         val cardNetworkResponses = googlePayManager.getSupportedCardNetworks(cardNetworks)
-        val googlePayEnabled =
-            when {
-                cardNetworkResponses.any { it.value === IsReadyToPayResponse.TRUE } ->
-                    IsReadyToPayResponse.TRUE
-                cardNetworkResponses.any { it.value === IsReadyToPayResponse.UNKNOWN } ->
-                    IsReadyToPayResponse.UNKNOWN
-                else -> IsReadyToPayResponse.FALSE
-            }
 
-        put("googlePayEnabled", googlePayEnabled)
+        put("googlePayEnabled", cardNetworkResponses.isGooglePlayEnabled())
         cardNetworkResponses.forEach { (network, supportResponse) ->
             put(
                 "googlePayCard_${network.value}",
